+++ car-showroom-NuxtJS/nuxt.config.ts
@@ -4,7 +4,7 @@ export default defineNuxtConfig({
   devtools: { enabled: true },
   components: true,
 
-  modules: ["@pinia/nuxt", "@vee-validate/nuxt"],
+  modules: ["@pinia/nuxt", "@vee-validate/nuxt","@vueuse/nuxt"],
   veeValidate: {
     autoImports: true,
     componentNames: {
