+++ car-showroom-NuxtJS/package.json
@@ -17,6 +17,10 @@
     "@pinia/nuxt": "^0.4.11",
     "@vee-validate/nuxt": "^4.9.6",
     "@vee-validate/rules": "^4.9.6",
+    "@vueuse/nuxt": "^10.1.2",
+    "axios": "^1.4.0",
+    "bootstrap": "^5.3.0",
+    "bootstrap-vue": "^2.23.1",
     "nuxt-storage": "^1.2.2",
     "vee-validate": "^4.9.6"
   }
